backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
apscheduler==3.10.4
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
tenacity==8.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0